+++ src/fmt.mli
@@ -86,6 +86,9 @@ val sp : unit t
 val comma : unit t
 (** [comma] is {!Fmt.unit}[ ",@ "]. *)
 
+val semi : unit t
+(** [semi] is {!Fmt.unit}[ ";@ "]. *)
+
 val const : 'a t -> 'a -> 'b t
 (** [const pp_v v] always formats [v] using [pp_v]. *)