+++ CHANGES.md
@@ -1,4 +1,5 @@
 
+* Add `Fmt.semi` for semicolon separators.
 * `Fmt.const` generalize signature.
 * Change signature of `Fmt.append`. Use `Fmt.(pair ~sep:nop)` if you
   were using it (backward compatible with earlier versions of `Fmt`).